+++ GameEntity/Tank.cpp
@@ -3,7 +3,7 @@
 /* The Constructor of Tank Object */
 Tank::Tank(
         const double& health, const double& health_regen, const double& max_health, const int& size,
-        const int& vx, const int& vy,const double& xp,
+        const double& vx, const double& vy,const double& xp,
         const double& attack_speed,
         const double& bullet_speed,
         const double& damage,
